@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "8.14.1",
+  "version": "8.14.2",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -29,8 +29,8 @@
     "sift": "17.1.3"
   },
   "devDependencies": {
-    "@babel/core": "7.26.10",
-    "@babel/preset-env": "7.26.9",
+    "@babel/core": "7.27.1",
+    "@babel/preset-env": "7.27.1",
     "@typescript-eslint/eslint-plugin": "^8.19.1",
     "@typescript-eslint/parser": "^8.19.1",
     "acquit": "1.3.0",
@@ -53,9 +53,9 @@
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
     "markdownlint-cli2": "^0.17.1",
-    "marked": "15.0.7",
+    "marked": "15.0.11",
     "mkdirp": "^3.0.1",
-    "mocha": "11.1.0",
+    "mocha": "11.2.2",
     "moment": "2.30.1",
     "mongodb-memory-server": "10.1.4",
     "ncp": "^2.0.0",
@@ -64,10 +64,10 @@
     "q": "1.5.1",
     "sinon": "20.0.0",
     "stream-browserify": "3.0.0",
-    "tsd": "0.31.2",
-    "typescript": "5.7.3",
+    "tsd": "0.32.0",
+    "typescript": "5.8.3",
     "uuid": "11.1.0",
-    "webpack": "5.98.0"
+    "webpack": "5.99.7"
   },
   "directories": {
     "lib": "./lib/mongoose"
@@ -101,7 +101,7 @@
     "mongo": "node ./tools/repl.js",
     "publish-7x": "npm publish --tag 7x",
     "test": "mocha --exit ./test/*.test.js",
-    "test-deno": "deno run --allow-env --allow-read --allow-net --allow-run --allow-sys --allow-write ./test/deno.js",
+    "test-deno": "deno run --allow-env --allow-read --allow-net --allow-run --allow-sys --allow-write ./test/deno.mjs",
     "test-rs": "START_REPLICA_SET=1 mocha --timeout 30000 --exit ./test/*.test.js",
     "test-tsd": "node ./test/types/check-types-filename && tsd",
     "setup-test-encryption": "bash scripts/configure-cluster-with-encryption.sh",
